缓存服务部分功能增强

+++ cache/src/main/java/com/bw/fit/common/cache/controller/CacheController.java
@@ -7,6 +7,7 @@ import com.bw.fit.common.cache.service.CacheService;
 import com.bw.fit.common.cache.util.PubFun;
 import com.google.common.collect.Iterators;
 import com.google.common.collect.Lists;
+import oracle.jdbc.proxy.annotation.Post;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
 import org.springframework.stereotype.Controller;
@@ -33,7 +34,7 @@ public class CacheController {
         return  (String)object;
     }
 
-    @PostMapping("setValue")
+    @PostMapping("cache")
     @ResponseBody
     public JSONObject setValue(@RequestParam(value = "key",required = true) String key,
                                @RequestParam(value = "value",required = true) String value){
@@ -48,16 +49,31 @@ public class CacheController {
         return json;
     }
 
-    @DeleteMapping("delete/{key}")
+    @DeleteMapping("cache/{key}")
     @ResponseBody
-    public JSONObject delete(@RequestParam(value = "key",required = true) String key ){
+    public JSONObject delete(@PathVariable(value = "key") String key ){
         JSONObject json = new JSONObject();
+        Object object1 = cacheService.get(key);
+        if (object1 == null){
+            PubFun.returnFailJson(json,"无效key");
+            return json;
+        }
         cacheService.delete(key);
         Object object = cacheService.get(key);
         if(object != null){
             PubFun.returnFailJson(json,"缓存操作失败,原因未知");
             return json;
         }
+        PubFun.returnSuccessJson(json);
+        return json;
+    }
+
+
+    @PostMapping("expire/{key}/{time}")
+    @ResponseBody
+    public JSONObject expire(@PathVariable String key,@PathVariable int time){
+        JSONObject json = new JSONObject();
+        cacheService.expire(key,time);
         PubFun.returnSuccessJson(json);
         return json;
     }
